compiled and updated the ethers-demo with package changes
