+++ packages/apps/ethers-demo/src/ethereum/contracts/Migrations.json
@@ -1391,10 +1391,22 @@
       "links": {},
       "address": "0x1B31aC789d12efA2244b508c5450a3350CA8e7aC",
       "transactionHash": "0x95050bc4cec82e0de8e2025ecc3f8e70017dddfbc9394c7f90b68188443f280a"
+    },
+    "1572631871774": {
+      "events": {},
+      "links": {},
+      "address": "0x96Af376b3557df65C20FE4c3251738787D5dF2AE",
+      "transactionHash": "0x2163953969b5a5381057dc6e6a81b3f3253eb7837d351e6de7b3ff4074c8ab43"
+    },
+    "1572891645705": {
+      "events": {},
+      "links": {},
+      "address": "0xB749d012Dadeb32f66077e68c1382c95c0d567F5",
+      "transactionHash": "0x02c251d5173ef0fe8156ae597bc23e6d1715100d102b569a761988483693c4d8"
     }
   },
   "schemaVersion": "3.0.16",
-  "updatedAt": "2019-10-31T16:15:03.772Z",
+  "updatedAt": "2019-11-04T20:17:08.014Z",
   "devdoc": {
     "methods": {}
   },
